28. Dagger внедрение (для списка пользователей)
