+++ app/src/main/res/layout/repository_fragment.xml
@@ -3,7 +3,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-   <!-- <LinearLayout
+   <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="8dp"
@@ -22,7 +22,7 @@
             android:layout_height="wrap_content"
             android:text="@string/User_name" />
 
-    </LinearLayout>-->
+    </LinearLayout>
 
     <FrameLayout
         android:layout_width="match_parent"
