+++ app/src/main/res/layout/user_list_fragment.xml
@@ -3,6 +3,26 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="8dp"
+        android:orientation="horizontal">
+
+        <EditText
+            android:id="@+id/enter_add_text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="4"
+            android:text="@string/any"/>
+
+        <Button
+            android:id="@+id/button_enter_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/User_name" />
+
+    </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
